# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My POS App API",
        "description": "API documentation for My POS App backend.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/sale-payments/{id}": {
            "get": {
                "tags": [
                    "SalePayment"
                ],
                "summary": "Get a single sale payment",
                "operationId": "310d505fdb96801f4359e93fda66aa3f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Sale payment not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "SalePayment"
                ],
                "summary": "Update a sale payment",
                "operationId": "7578ab25db588833ea36e24712f0457a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sale payment updated"
                    },
                    "404": {
                        "description": "Sale payment not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "SalePayment"
                ],
                "summary": "Delete a sale payment",
                "operationId": "8832d53c37ae146f800a5aac7a591c04",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Sale payment deleted"
                    },
                    "404": {
                        "description": "Sale payment not found"
                    }
                }
            }
        },
        "/sale-payments": {
            "get": {
                "tags": [
                    "SalePayment"
                ],
                "summary": "List all sale payments",
                "operationId": "dc66e5cbcc15f213bcfffd7c9741ea7a",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "SalePayment"
                ],
                "summary": "Create a new sale payment",
                "operationId": "3d807c00279de068fc51ccb8189608ed",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sale payment created"
                    }
                }
            }
        },
        "/shops": {
            "get": {
                "tags": [
                    "Shop"
                ],
                "summary": "List all shops",
                "operationId": "4a5ba44aff1934ce0f1cb0c335dba1a4",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Shop"
                ],
                "summary": "Create a new shop",
                "operationId": "c7a906bde8b73fa4690a9f768ac6db19",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Shop created"
                    }
                }
            }
        },
        "/shops/{id}": {
            "get": {
                "tags": [
                    "Shop"
                ],
                "summary": "Get a single shop",
                "operationId": "85bbc335740ec8946003594d34996d01",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Shop not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Shop"
                ],
                "summary": "Update a shop",
                "operationId": "10e9239d79889df46fcb1c2e103f386e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Shop updated"
                    },
                    "404": {
                        "description": "Shop not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Shop"
                ],
                "summary": "Delete a shop",
                "operationId": "1dfea7acd7d35abd6f3cd0a929d23fe0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Shop deleted"
                    },
                    "404": {
                        "description": "Shop not found"
                    }
                }
            }
        },
        "/suppliers": {
            "get": {
                "tags": [
                    "Supplier"
                ],
                "summary": "List all suppliers",
                "operationId": "880dea28ea6be2c0748bfd5895726b27",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Supplier"
                ],
                "summary": "Create a new supplier",
                "operationId": "658eee9a3a5922bcc10d9e13b637c4b5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Supplier created"
                    }
                }
            }
        },
        "/suppliers/{id}": {
            "get": {
                "tags": [
                    "Supplier"
                ],
                "summary": "Get a single supplier",
                "operationId": "7d893aa4dc7a3c49058854f71b35ac3b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Supplier not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Supplier"
                ],
                "summary": "Update a supplier",
                "operationId": "6092f70b7812608de6dd450a9eb0ae2a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Supplier updated"
                    },
                    "404": {
                        "description": "Supplier not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Supplier"
                ],
                "summary": "Delete a supplier",
                "operationId": "f93c2db0ee9f314d86463df419ae8099",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Supplier deleted"
                    },
                    "404": {
                        "description": "Supplier not found"
                    }
                }
            }
        },
...
    },
...
}